requirements.txt
fastapi
uvicorn
python-multipart
google-genai
python-dotenv
pillow
numpy
sentence-transformers
pypdf
httpx
streamlit
requests
